requirements.txt
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.20.0